Using a fill pattern much larger than actual canvas reliably segfaults browser https://bugs.webkit.org/show_bug.cgi?id=132635
Reviewed by Simon Fraser.

Source/WebCore:
Make sure that createPattern checks that the canvas it is about to use
as a source is valid.

Test: fast/canvas/pattern-too-large-to-create.html

* html/canvas/CanvasRenderingContext2D.cpp:
(WebCore::CanvasRenderingContext2D::createPattern): Check that the source canvas has
an ok ImageBuffer.

LayoutTests:
Test case that makes a huge canvas and tries to create a pattern out of it.

* fast/canvas/pattern-too-large-to-create-expected.html: Added.
* fast/canvas/pattern-too-large-to-create.html: Added.

Identifier: 150719@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@168457 268f45cc-cd09-0410-ab3c-d52691b4dbfc