[CSS Regions] Add ASSERT to make sure using the flowThread cache does not return incorrect results https://bugs.webkit.org/show_bug.cgi?id=132906
Reviewed by Andrei Bucur.

If flowThreadContainingBlock() is called on an object which is in a different
flow thread than the one currently being laid out, this method will return an incorrect
result. I added an assertion for that to make sure we catch and treat any such scenarios.

No new tests required.

* rendering/RenderObject.cpp:
(WebCore::RenderObject::locateFlowThreadContainingBlockNoCache):
(WebCore::RenderObject::locateFlowThreadContainingBlock):
* rendering/RenderObject.h:


Identifier: 150963@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@168837 268f45cc-cd09-0410-ab3c-d52691b4dbfc